add compose

+++ test/test_helper.exs
@@ -4,19 +4,24 @@ File.mkdir("tmp")
 Mox.defmock(Hexdocs.HexpmMock, for: Hexdocs.Hexpm)
 Mox.defmock(Hexdocs.SourceRepo.Mock, for: Hexdocs.SourceRepo)
 
-exclude =
-  case Hexdocs.HTTP.get("http://localhost:8108/health", _req_headers = []) do
-    {:ok, 200, _resp_headers, ~s|{"ok":true}|} ->
-      _no_exclude = []
+if :typesense in ExUnit.configuration()[:include] do
+  typesense_available? =
+    case Hexdocs.HTTP.get("http://localhost:8108/health", _req_headers = []) do
+      {:ok, 200, _resp_headers, ~s|{"ok":true}|} -> true
+      {:error, :econnrefused} -> false
+    end
 
-    {:error, :econnrefused} ->
-      Mix.shell().error("""
-      To enable Typesense tests, start the local container with the following command:
+  unless typesense_available? do
+    Mix.shell().error("""
+    To enable Typesense tests, start the local container with the following command:
 
-          docker run -d --rm -p 8108:8108 typesense/typesense:26.0 --data-dir /tmp --api-key=hexdocs
-      """)
+        docker compose up -d
 
-      _exclude_typesense = [:typesense]
+    Or, if you prefer to run the container manually:
+
+        docker run -d --rm -p 8108:8108 typesense/typesense:26.0 --data-dir /tmp --api-key=hexdocs
+    """)
   end
+end
 
-ExUnit.start(exclude: exclude)
+ExUnit.start(exclude: [:typesense])
